Kal 10 youtube scraper

+++ backend/api/src/main/java/ai/kalico/api/service/youtubej/extractor/Extractor.java
@@ -20,6 +20,6 @@ public interface Extractor {
 
     String extractClientVersionFromContext(JSONObject context);
 
-    int extractIntegerFromText(String text);
+    long extractIntegerFromText(String text);
 
 }
